+++ controllers/products.js
@@ -3,6 +3,7 @@ const User = require('../models/user');
 const multer = require('multer')
 const fs = require('fs');
 const path = require('path');
+const Image = require('../models/images');
 
 // RENDER THE NEW PRODUCT FORM
 function newProd(req, res) {
@@ -11,29 +12,124 @@ function newProd(req, res) {
 
 
 /// POST THE DETAILS OF NEW PRODUCT FORM TO HOME
-function create(req, res) {
-    console.log(req.body)
-    let newProd = new Products({
-        title: req.body.title,
-        shortdes: req.body.shortdes,
-        longdes: req.body.longdes,
-        seller: req.user,
-        image: {
-            
-        }
-    })
+// function create(req, res) {
 
-    console.log("NEWPROD" + newProd)
-    newProd.save()
+//     let newProd = new Products({
+//         title: req.body.title,
+//         shortdes: req.body.shortdes,
+//         longdes: req.body.longdes,
+//         seller: req.user,
+//         image: null,
+//     })
 
-    User.findById(newProd.seller).exec(function (err, foundUser) {
-        if (err) res.send(err);
-        foundUser.products.push(newProd._id);
-        foundUser.save();
+//     console.log("NEWPROD" + newProd)
+//     newProd.save()
+    
+//     User.findById(newProd.seller).exec(function (err, foundUser) {
+//         if (err) res.send(err);
+//         foundUser.products.push(newProd._id);
+//         foundUser.save();
 
-    });
-    res.redirect('/')
+//     });
+//     res.redirect('/')
+// }
+
+
+/* ====== IMAGE UPLOAD  ====== */ 
+// Set The Storage Engine
+const Storage = multer.diskStorage({
+  destination: './public/uploads',
+  filename: function(req, file, cb){
+    cb(null,file.originalname);
+  }
+});
+ 
+
+// Check File Type
+function checkFileType(file, cb){
+  // Allowed ext
+  const filetypes = /jpeg|jpg|png|gif/;
+  // Check ext
+  const extname = filetypes.test(path.extname(file.originalname).toLowerCase());
+  // Check mime
+  const mimetype = filetypes.test(file.mimetype);
+
+  if(mimetype && extname){
+    return cb(null,true);
+  } else {
+    cb('Error: Images Only!');
+  }
 }
+
+  // Init Upload
+  const upload = multer({
+    storage: Storage,
+    limits:{fileSize: 1000000},
+    fileFilter: function(req, file, cb){
+      checkFileType(file, cb);
+    }
+  })
+
+function create(req, res) {
+  upload.single('image')(req, res, function(err) {
+    if (err) {
+      console.log(err);
+      return res.status(500).send({ message: 'Error uploading image.' });
+    }
+
+    let newProd = new Products({
+      title: req.body.title,
+      shortdes: req.body.shortdes,
+      longdes: req.body.longdes,
+      seller: req.user,
+      image: null
+    });
+
+    newProd.save(function(err, prod) {
+      if (err) {
+        console.log(err);
+        return res.status(500).send({ message: 'Error creating product.' });
+      }
+
+      if (req.file) {
+        let newImage = new Image({
+          name: req.file.originalname,
+          image: {
+            data: req.file.buffer,
+            contentType: req.file.mimetype
+          },
+          filepath: req.file.path
+        });
+
+        newImage.save(function(err, image) {
+          if (err) {
+            console.log(err);
+            return res.status(500).send({ message: 'Error creating image.' });
+          }
+
+          Products.findByIdAndUpdate(newProd._id, { image: image._id }, function(err) {
+            if (err) {
+              console.log(err);
+              return res.status(500).send({ message: 'Error updating product with image.' });
+            }
+
+            Products.findById(newProd._id).populate('image').exec(function(err, product) {
+              if (err) {
+                console.log(err);
+                return res.status(500).send({ message: 'Error populating image into product.' });
+              }
+
+              res.redirect('/');
+            });
+          });
+        });
+      } else {
+        res.redirect('/');
+      }
+    });
+  });
+}
+
 
 // RENDER THE PRODUCT ID PAGE
 function prodId(req, res) {
